Remove usage of annotations from tests
