+++ tests/Functional/Controller/TestController.php
@@ -11,37 +11,15 @@
 
 namespace Nelmio\ApiDocBundle\Tests\Functional\Controller;
 
-use Nelmio\ApiDocBundle\Tests\Functional\TestKernel;
-use OpenApi\Annotations as OA;
-use OpenApi\Attributes as OAT;
+use OpenApi\Attributes as OA;
 use Symfony\Component\Routing\Annotation\Route;
 
-if (TestKernel::isAnnotationsAvailable()) {
-    /**
-     * @Route("/test", host="api-test.example.com")
-     */
-    class TestController
+#[Route('/test', host: 'api-test.example.com')]
+class TestController
+{
+    #[OA\Response(response: 200, description: 'Test')]
+    #[Route('/test/', methods: ['GET'])]
+    public function testAction(): void
     {
-        /**
-         * @OA\Response(
-         *     response="200",
-         *     description="Test"
-         * )
-         *
-         * @Route("/test/", methods={"GET"})
-         */
-        public function testAction(): void
-        {
-        }
-    }
-} else {
-    #[Route('/test', host: 'api-test.example.com')]
-    class TestController
-    {
-        #[OAT\Response(response: 200, description: 'Test')]
-        #[Route('/test/', methods: ['GET'])]
-        public function testAction(): void
-        {
-        }
     }
 }
