+++ tests/Functional/Entity/Article.php
@@ -11,36 +11,16 @@
 
 namespace Nelmio\ApiDocBundle\Tests\Functional\Entity;
 
-use Nelmio\ApiDocBundle\Tests\Functional\TestKernel;
 use Symfony\Component\Serializer\Annotation\Groups;
 
-if (TestKernel::isAnnotationsAvailable()) {
-    /**
-     * @author Guilhem N. <guilhem.niot@gmail.com>
-     */
-    class Article
+class Article
+{
+    #[Groups(['light'])]
+    public function setAuthor(User $author)
     {
-        /**
-         * @Groups({"light"})
-         */
-        public function setAuthor(User $author)
-        {
-        }
-
-        public function setContent(string $content)
-        {
-        }
     }
-} else {
-    class Article
-    {
-        #[Groups(['light'])]
-        public function setAuthor(User $author)
-        {
-        }
 
-        public function setContent(string $content)
-        {
-        }
+    public function setContent(string $content)
+    {
     }
 }
