--- tests/Functional/Entity/NestedGroup/JMSChat80.php
@@ -1,36 +1,0 @@
-<?php
-
-/*
- * This file is part of the NelmioApiDocBundle package.
- *
- * (c) Nelmio
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace Nelmio\ApiDocBundle\Tests\Functional\Entity\NestedGroup;
-
-use JMS\Serializer\Annotation as Serializer;
-
-/**
- * User.
- *
- * @Serializer\ExclusionPolicy("all")
- */
-class JMSChat80
-{
-    /**
-     * @Serializer\Type("integer")
-     *
-     * @Serializer\Expose
-     */
-    private $id;
-
-    /**
-     * @Serializer\Type("array<Nelmio\ApiDocBundle\Tests\Functional\Entity\NestedGroup\JMSChatUser>")
-     *
-     * @Serializer\Expose
-     */
-    private $members;
-}
