+++ tests/Functional/Entity/NestedGroup/JMSChatRoom.php
@@ -11,14 +11,25 @@
 
 namespace Nelmio\ApiDocBundle\Tests\Functional\Entity\NestedGroup;
 
-use Nelmio\ApiDocBundle\Tests\Functional\TestKernel;
+use JMS\Serializer\Annotation as Serializer;
 
-if (TestKernel::isAnnotationsAvailable()) {
-    class JMSChatRoom extends JMSChatRoom80
-    {
-    }
-} else {
-    class JMSChatRoom extends JMSChatRoom81
-    {
-    }
+/**
+ * User.
+ */
+#[Serializer\ExclusionPolicy('all')]
+class JMSChatRoom
+{
+    #[Serializer\Type('integer')]
+    #[Serializer\Expose]
+    #[Serializer\Groups(['Default'])]
+    private $id1;
+
+    #[Serializer\Type('integer')]
+    #[Serializer\Expose]
+    #[Serializer\Groups(['mini'])]
+    private $id2;
+
+    #[Serializer\Type('integer')]
+    #[Serializer\Expose]
+    private $id3;
 }
