+++ tests/Functional/Entity/NestedGroup/JMSChatRoomUser.php
@@ -11,14 +11,17 @@
 
 namespace Nelmio\ApiDocBundle\Tests\Functional\Entity\NestedGroup;
 
-use Nelmio\ApiDocBundle\Tests\Functional\TestKernel;
+use JMS\Serializer\Annotation as Serializer;
 
-if (TestKernel::isAnnotationsAvailable()) {
-    class JMSChatRoomUser extends JMSChatRoomUser80
-    {
-    }
-} else {
-    class JMSChatRoomUser extends JMSChatRoomUser81
-    {
-    }
+#[Serializer\ExclusionPolicy('all')]
+class JMSChatRoomUser
+{
+    #[Serializer\Type('integer')]
+    #[Serializer\Expose]
+    private $id;
+
+    #[Serializer\Type('Nelmio\ApiDocBundle\Tests\Functional\Entity\NestedGroup\JMSChatFriend')]
+    #[Serializer\Expose]
+    #[Serializer\Groups(['mini'])]
+    private $friend;
 }
