--- tests/Functional/Entity/NestedGroup/JMSChatRoomUser81.php
@@ -1,27 +1,0 @@
-<?php
-
-/*
- * This file is part of the NelmioApiDocBundle package.
- *
- * (c) Nelmio
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace Nelmio\ApiDocBundle\Tests\Functional\Entity\NestedGroup;
-
-use JMS\Serializer\Annotation as Serializer;
-
-#[Serializer\ExclusionPolicy('all')]
-class JMSChatRoomUser81
-{
-    #[Serializer\Type('integer')]
-    #[Serializer\Expose]
-    private $id;
-
-    #[Serializer\Type('Nelmio\ApiDocBundle\Tests\Functional\Entity\NestedGroup\JMSChatFriend')]
-    #[Serializer\Expose]
-    #[Serializer\Groups(['mini'])]
-    private $friend;
-}
