+++ tests/Functional/Entity/User.php
@@ -11,303 +11,144 @@
 
 namespace Nelmio\ApiDocBundle\Tests\Functional\Entity;
 
-use Nelmio\ApiDocBundle\Tests\Functional\TestKernel;
-use OpenApi\Annotations as OA;
+use OpenApi\Attributes as OA;
 
-if (TestKernel::isAnnotationsAvailable()) {
+/**
+ * @author Guilhem N. <egetick@gmail.com>
+ */
+class User
+{
     /**
-     * @author Guilhem N. <egetick@gmail.com>
+     * @var int
      */
-    class User
+    #[OA\Property(description: 'User id', readOnly: true, title: 'userid', default: null)]
+    private $id;
+
+    #[OA\Property(type: 'string', readOnly: false)]
+    private $email;
+
+    /**
+     * User Roles Comment.
+     *
+     * @var string[]
+     */
+    #[OA\Property(description: 'User roles', title: 'roles', example: '["ADMIN","SUPERUSER"]', default: ['user'])]
+    private $roles;
+
+    /**
+     * User Location.
+     */
+    #[OA\Property(type: 'string')]
+    private $location;
+
+    /**
+     * @var int
+     */
+    #[OA\Property(type: 'string')]
+    private $friendsNumber;
+
+    /**
+     * @var float
+     */
+    #[OA\Property(default: 0.0)]
+    private $money;
+
+    /**
+     * @var \DateTime
+     */
+    #[OA\Property(property: 'creationDate')]
+    private $createdAt;
+
+    /**
+     * @var User[]
+     */
+    private $users;
+
+    /**
+     * @var User|null
+     */
+    private $friend;
+
+    /**
+     * @var User[]|null
+     */
+    private $friends;
+
+    /**
+     * @var string
+     */
+    #[OA\Property(enum: ['disabled', 'enabled'])]
+    private $status;
+
+    /**
+     * @var \DateTimeInterface
+     */
+    private $dateAsInterface;
+
+    public function setMoney(float $money)
     {
-        /**
-         * @var int
-         *
-         * @OA\Property(description = "User id", readOnly = true, title = "userid", default = null)
-         */
-        private $id;
-
-        /**
-         * @OA\Property(type="string", readOnly = false)
-         */
-        private $email;
-
-        /**
-         * User Roles Comment.
-         *
-         * @var string[]
-         *
-         * @OA\Property(
-         *     description = "User roles",
-         *     title = "roles",
-         *     example="[""ADMIN"",""SUPERUSER""]",
-         *     default = {"user"},
-         * )
-         */
-        private $roles;
-
-        /**
-         * User Location.
-         *
-         * @OA\Property(type = "string")
-         */
-        private $location;
-
-        /**
-         * @var int
-         *
-         * @OA\Property(type = "string")
-         */
-        private $friendsNumber;
-
-        /**
-         * @var float
-         *
-         * @OA\Property(default = 0.0)
-         */
-        private $money;
-
-        /**
-         * @var \DateTime
-         *
-         * @OA\Property(property="creationDate")
-         */
-        private $createdAt;
-
-        /**
-         * @var User[]
-         */
-        private $users;
-
-        /**
-         * @var User|null
-         */
-        private $friend;
-
-        /**
-         * @var User[]|null
-         */
-        private $friends;
-
-        /**
-         * @var string
-         *
-         * @OA\Property(enum = {"disabled", "enabled"})
-         */
-        private $status;
-
-        /**
-         * @var \DateTimeInterface
-         */
-        private $dateAsInterface;
-
-        public function setMoney(float $money)
-        {
-            $this->money = $money;
-        }
-
-        /**
-         * @OA\Property(example=1)
-         */
-        public function setId(int $id)
-        {
-            $this->id = $id;
-        }
-
-        public function setEmail($email)
-        {
-            $this->email = $email;
-        }
-
-        /**
-         * @param string[] $roles
-         */
-        public function setRoles(array $roles)
-        {
-            $this->roles = $roles;
-        }
-
-        public function setLocation(string $location)
-        {
-        }
-
-        public function setFriendsNumber(int $friendsNumber)
-        {
-            $this->friendsNumber = $friendsNumber;
-        }
-
-        public function setCreatedAt(\DateTime $createAt)
-        {
-        }
-
-        public function setUsers(array $users)
-        {
-        }
-
-        public function setFriend(?self $friend = null)
-        {
-        }
-
-        public function setFriends(array $friends = [])
-        {
-        }
-
-        public function setDummy(Dummy $dummy)
-        {
-        }
-
-        public function setStatus(string $status)
-        {
-        }
-
-        public function getDateAsInterface(): \DateTimeInterface
-        {
-            return $this->dateAsInterface;
-        }
-
-        public function setDateAsInterface(\DateTimeInterface $dateAsInterface)
-        {
-            $this->dateAsInterface = $dateAsInterface;
-        }
+        $this->money = $money;
     }
-} else {
-    /**
-     * @author Guilhem N. <egetick@gmail.com>
-     */
-    class User
+
+    #[OA\Property(example: 1)]
+    public function setId(int $id)
     {
-        /**
-         * @var int
-         */
-        #[\OpenApi\Attributes\Property(description: 'User id', readOnly: true, title: 'userid', default: null)]
-        private $id;
+        $this->id = $id;
+    }
 
-        #[\OpenApi\Attributes\Property(type: 'string', readOnly: false)]
-        private $email;
+    public function setEmail($email)
+    {
+        $this->email = $email;
+    }
 
-        /**
-         * User Roles Comment.
-         *
-         * @var string[]
-         */
-        #[\OpenApi\Attributes\Property(description: 'User roles', title: 'roles', example: '["ADMIN","SUPERUSER"]', default: ['user'])]
-        private $roles;
+    /**
+     * @param string[] $roles
+     */
+    public function setRoles(array $roles)
+    {
+        $this->roles = $roles;
+    }
 
-        /**
-         * User Location.
-         */
-        #[\OpenApi\Attributes\Property(type: 'string')]
-        private $location;
+    public function setLocation(string $location)
+    {
+    }
 
-        /**
-         * @var int
-         */
-        #[\OpenApi\Attributes\Property(type: 'string')]
-        private $friendsNumber;
+    public function setFriendsNumber(int $friendsNumber)
+    {
+        $this->friendsNumber = $friendsNumber;
+    }
 
-        /**
-         * @var float
-         */
-        #[\OpenApi\Attributes\Property(default: 0.0)]
-        private $money;
+    public function setCreatedAt(\DateTime $createAt)
+    {
+    }
 
-        /**
-         * @var \DateTime
-         */
-        #[\OpenApi\Attributes\Property(property: 'creationDate')]
-        private $createdAt;
+    public function setUsers(array $users)
+    {
+    }
 
-        /**
-         * @var User[]
-         */
-        private $users;
+    public function setFriend(?self $friend = null)
+    {
+    }
 
-        /**
-         * @var User|null
-         */
-        private $friend;
+    public function setFriends(array $friends = [])
+    {
+    }
 
-        /**
-         * @var User[]|null
-         */
-        private $friends;
+    public function setDummy(Dummy $dummy)
+    {
+    }
 
-        /**
-         * @var string
-         */
-        #[\OpenApi\Attributes\Property(enum: ['disabled', 'enabled'])]
-        private $status;
+    public function setStatus(string $status)
+    {
+    }
 
-        /**
-         * @var \DateTimeInterface
-         */
-        private $dateAsInterface;
+    public function getDateAsInterface(): \DateTimeInterface
+    {
+        return $this->dateAsInterface;
+    }
 
-        public function setMoney(float $money)
-        {
-            $this->money = $money;
-        }
-
-        #[\OpenApi\Attributes\Property(example: 1)]
-        public function setId(int $id)
-        {
-            $this->id = $id;
-        }
-
-        public function setEmail($email)
-        {
-            $this->email = $email;
-        }
-
-        /**
-         * @param string[] $roles
-         */
-        public function setRoles(array $roles)
-        {
-            $this->roles = $roles;
-        }
-
-        public function setLocation(string $location)
-        {
-        }
-
-        public function setFriendsNumber(int $friendsNumber)
-        {
-            $this->friendsNumber = $friendsNumber;
-        }
-
-        public function setCreatedAt(\DateTime $createAt)
-        {
-        }
-
-        public function setUsers(array $users)
-        {
-        }
-
-        public function setFriend(?self $friend = null)
-        {
-        }
-
-        public function setFriends(array $friends = [])
-        {
-        }
-
-        public function setDummy(Dummy $dummy)
-        {
-        }
-
-        public function setStatus(string $status)
-        {
-        }
-
-        public function getDateAsInterface(): \DateTimeInterface
-        {
-            return $this->dateAsInterface;
-        }
-
-        public function setDateAsInterface(\DateTimeInterface $dateAsInterface)
-        {
-            $this->dateAsInterface = $dateAsInterface;
-        }
+    public function setDateAsInterface(\DateTimeInterface $dateAsInterface)
+    {
+        $this->dateAsInterface = $dateAsInterface;
     }
 }
